MDEV-32514 Slow log entries causing main.subselect mtr failures
In MDEV-7445, a test required an empty table as a data source.
mysql.slow_log was chosen as this table, but some CI testers
enable this in a worker as a command line option, meaning that
this table may no longer be empty. If the table returns more
than 1 row, the query fails, causing test failure.
This will clearly be a random failure, depending on OS workload.
Removed mysql.slow_log as data source.